% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygraph-funs.R
\name{as_tidygraph_list}
\alias{as_tidygraph_list}
\title{Get posterior tree draws into \code{tbl_graph} format from \code{tidygraph} package.}
\usage{
as_tidygraph_list(model, extra_cols, label_digits = 2)
}
\arguments{
\item{model}{BART model.}

\item{extra_cols}{Extra columns to be included.}

\item{label_digits}{Rounding for labels.}
}
\value{
List of \code{tbl_graph}.
}
\description{
Creates a list of \code{tbl_graph}'s. Each element of list corresponds to a particular MCMC iteration (\code{iter}) and tree id (\code{tree_id}).
See \link[tidygraph:tbl_graph]{tbl_graph} for details.
}
