% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytree-funs.R
\name{as_tidytree}
\alias{as_tidytree}
\title{Get posterior tree draws into \code{tbl_tree} format from \code{tidytree} package.}
\usage{
as_tidytree(model, extra_cols, label_digits = 2)
}
\arguments{
\item{model}{BART model.}

\item{extra_cols}{Extra columns to be included.}

\item{label_digits}{Rounding for labels.}
}
\value{
A tibble with required columns for \code{tbl_tree}:
\code{node}, \code{parent}, \code{label}.
And columns \code{iter}, \code{tree_id} are used to differentiate trees and
iterations and are always included. Remaining columns are optional (see
details).
}
\description{
Creates a \code{tbl_tree} grouped by iteration (\code{iter}) and tree id (\code{tree_id}).
See Details below.
}
\details{
List of potential columns returned:
\describe{
\item{iter}{Integer describing unique MCMC iteration.}
\item{tree_id}{Integer. Unique tree id with each \code{iter}.}
\item{node}{Integer describing node in tree. Unique to each \code{tree}-\code{iter}.}
\item{parent}{Integer describing parent node in tree.}
\item{label}{Label for the node.}
\item{tier}{Position in tree hierarchy.}
\item{var}{Variable for split.}
\item{cut}{Numeric. Value of decision rule for \code{var}.}
\item{is_leaf}{Logical. \code{TRUE} if leaf, \code{FALSE} if stem.}
\item{leaf_value}{Numeric (mean function) value of leaf}
\item{child_left}{Integer. Left child of node.}
\item{child_right}{Integer. Right child of node.}
}
}
