% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eatmyshorts-package.R
\docType{package}
\name{eatmyshorts}
\alias{eatmyshorts}
\title{eatmyshorts: Visualisation for BART models}
\description{
Visualisation methods and functions for exploring Bayesian Additive
Regression Trees
}
