% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BART-graph.R
\name{posterior_trees}
\alias{posterior_trees}
\title{Get posterior tree draws into tibble format}
\usage{
posterior_trees(model)
}
\arguments{
\item{model}{BART model}
}
\value{
A tibble
}
\description{
Tibble grouped by iteration (\code{iter}) and tree id (\code{tree_id}).
}
