% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BART-graph.R
\name{posterior_trees}
\alias{posterior_trees}
\title{Get posterior tree draws into XXX format}
\usage{
posterior_trees(model)
}
\arguments{
\item{model}{BART model}
}
\value{
List of XXX objects.
}
\description{
Get posterior tree draws into XXX format
}
